requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dateutil>=2.8.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
google-genai>=0.2.0
pyaudio>=0.2.11
numpy>=1.24.0

